move some code from core to physical_expr crate
